requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.42.0
python-telegram-bot==22.7
python-dotenv==1.2.2
httpx==0.28.1
pytest==9.0.2
pytest-asyncio==1.3.0
